fix: fix cyber build

+++ cyber/logger/logger.cc
@@ -45,7 +45,7 @@ Logger::~Logger() {
 void Logger::Write(bool force_flush,
                    const std::chrono::system_clock::time_point& timestamp,
                    const char* message,
-                   int message_len) {
+                   size_t message_len) {
   std::string log_message = std::string(message, message_len);
   std::string module_name;
   // set the same bracket as the bracket in log.h
